fix compiler warning -Wimplicit-int-float-conversion
just use floats for all si math since int to float conversion
would convert 1000000000000 to 999999995904

+++ src/vfs/vfs-file-info.c
@@ -443,7 +443,7 @@ void vfs_file_size_to_string_format(char* buf, uint64_t size, char* format) // M
             */
             if (app_settings.use_si_prefix)
             {
-                val = ((float)size) / ((uint64_t)1000000000000);
+                val = ((float)size) / ((float)1000000000000);
                 unit = _("T");
             }
             else
@@ -461,7 +461,7 @@ void vfs_file_size_to_string_format(char* buf, uint64_t size, char* format) // M
             */
             if (app_settings.use_si_prefix)
             {
-                val = ((float)size) / ((uint64_t)1000000000);
+                val = ((float)size) / ((float)1000000000);
                 unit = _("G");
             }
             else
@@ -480,7 +480,7 @@ void vfs_file_size_to_string_format(char* buf, uint64_t size, char* format) // M
         */
         if (app_settings.use_si_prefix)
         {
-            val = ((float)size) / ((uint64_t)1000000);
+            val = ((float)size) / ((float)1000000);
             unit = _("M");
         }
         else
@@ -498,7 +498,7 @@ void vfs_file_size_to_string_format(char* buf, uint64_t size, char* format) // M
         */
         if (app_settings.use_si_prefix)
         {
-            val = ((float)size) / ((uint64_t)1000);
+            val = ((float)size) / ((float)1000);
             unit = _("K"); // MOD was KB
         }
         else
@@ -542,7 +542,7 @@ void vfs_file_size_to_string(char* buf, uint64_t size)
             */
             if (app_settings.use_si_prefix)
             {
-                val = ((float)size) / ((uint64_t)1000000000000);
+                val = ((float)size) / ((float)1000000000000);
                 unit = _("T");
             }
             else
@@ -560,7 +560,7 @@ void vfs_file_size_to_string(char* buf, uint64_t size)
             */
             if (app_settings.use_si_prefix)
             {
-                val = ((float)size) / ((uint64_t)1000000000);
+                val = ((float)size) / ((float)1000000000);
                 unit = _("G");
             }
             else
@@ -579,7 +579,7 @@ void vfs_file_size_to_string(char* buf, uint64_t size)
         */
         if (app_settings.use_si_prefix)
         {
-            val = ((float)size) / ((uint64_t)1000000);
+            val = ((float)size) / ((float)1000000);
             unit = _("M");
         }
         else
@@ -597,7 +597,7 @@ void vfs_file_size_to_string(char* buf, uint64_t size)
         */
         if (app_settings.use_si_prefix)
         {
-            val = ((float)size) / ((uint64_t)1000);
+            val = ((float)size) / ((float)1000);
             unit = _("K"); // MOD was KB
         }
         else
